Fix SonarCloud issues
- Circular dependency between locale.js and app_ui_settings.js (in components/frontend/src).
- Remove deprecated inputProps in components/frontend/src/widgets/buttons/AddDropdownButton.jsx.
- Prevent potential attribute error in components/shared_code/tests/shared/model/test_metric.py.
